pyproject.toml
[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "handwriting-ai"
version = "0.1.0"
description = "Handwriting AI service for MNIST digit inference (HTTP)."
authors = ["Your Team <team@example.com>"]
packages = [{ include = "handwriting_ai", from = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
Pillow = "^10.4.0"
numpy = ">=1.26,<3"
python-multipart = "^0.0.9"
typing-extensions = "^4.12.2"
# Torch CPU (install wheels as appropriate for platform)
torch = {version = "^2.4.1", markers = "platform_system != 'Darwin' or platform_machine != 'arm64'"}
torchvision = {version = "^0.19.1", markers = "platform_system != 'Darwin' or platform_machine != 'arm64'"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.2"
ruff = "^0.6.5"
black = "^24.10.0"
httpx = "^0.27.2"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
exclude = [".venv"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # bugbear
  "BLE",# blind except
  "UP", # pyupgrade
  "N",  # pep8-naming
  "C4", # flake8-comprehensions
  "SIM",# flake8-simplify
  "RET",# flake8-return
  "C90",# mccabe complexity
  "RUF",# ruff-specific
]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
no_implicit_optional = true
check_untyped_defs = true
no_implicit_reexport = true
show_error_codes = true
pretty = true
files = ["src", "tests"]
plugins = ["pydantic.mypy"]

# Narrow, explicit relaxations for third‑party integration boundaries to avoid false Any from
# external stubs, while keeping our codebase free of Any/casts/ignores.
 

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
precision = 2
show_missing = true
